fix cpu import and env vars
